Fix setting default status on task orders
